// src/constants/mutation-type.ts
export enum MutationType {
  SET_DATA_SERIES_ENTRY = 'setDataSeriesEntry',
  SET_DID = 'setDID',
  SET_DATA_URL = 'setDataUrl',
  SET_SCHEMA_DRI = 'setSchemaDri',
  SET_SURVEY = 'setSurvey',
  SET_SURVEY_META_ITEM = 'setSurveyMetaItem',
  SET_CONTROLLER_USAGE_POLICY = 'setControllerUsagePolicy',
  SET_IS_USAGE_POLICY_MATCHING = 'setIsUsagePolicyMatching',
  RESET_STATE = 'resetState',
}